Modified it to take command line arguments.

+++ install.sh
@@ -1,4 +1,4 @@
-#!/bin/sh
+#!/bin/zsh
 
 # Installing curl if not installed
 if which curl >/dev/null 2>&1; then
@@ -41,6 +41,25 @@ else
 fi
 
 # Downloading the oh-my-zsh theme (you can use your favorite theme)
+change_theme() {
+	local NEW_THEME=$1
+	local ZSHRC_FILE="$HOME/.zshrc"
+
+	sed -i "11s/^ZSH_THEME=\".*\"/ZSH_THEME=\"$NEW_THEME\"/" "$ZSHRC_FILE"
+	. "$ZSHRC_FILE"
+	echo "Updated the theme."
+	echo "*If the terminal doesn't look updated close and restart the shell.*"
+	exit 0
+}
+
+if [ "$1" = "--change-theme" ] && [ -n "$2" ]; then
+	echo "Enter the new zsh theme name\n"
+	echo "If you don't know the theme name read the README file of the git repo: "
+	THEME_NAME="$2"
+	change_theme "$THEME_NAME"
+	exit 0
+fi
+
 read -p "Enter the GitHub link for zsh theme: " GIT_LINK
 
 BASE_DIR="$HOME/.zsh-themes"
@@ -63,12 +82,26 @@ else
 	echo "Cloned '$THEME_NAME' into $TARGET_DIR"
 fi
 
-cp $TARGET_DIR/$THEME_NAME $ZSH/themes/$THEME_NAME
+if [ -f "$TARGET_DIR/$THEME_NAME" ]; then
+	if [ ! -f "$ZSH/themes/$THEME_NAME" ]; then
+		cp $TARGET_DIR/$THEME_NAME $ZSH/themes/$THEME_NAME
+		echo "Copied '$THEME_NAME' to '$ZSH/themes/'"
+	else
+		echo "\e[1mFile '$THEME_NAME' already exists in '$ZSH/themes/'. Skipping copy.\e[0m"
+	fi
+else
+	echo "The file '$THEME_NAME' doesn't exist in '$TARGET_DIR'"
+	exit 1
+fi
 
 ZSHRC_FILE="$HOME/.zshrc"
+
 echo "Enter the new zsh theme name\n"
 read -p "If you don't know the theme name read the README file of the git repo: " NEW_THEME
 
 sed -i "11s/^ZSH_THEME=\".*\"/ZSH_THEME=\"$NEW_THEME\"/" "$ZSHRC_FILE"
-source $HOME/.zshrc
+. "$ZSHRC_FILE"
 echo "Updated the theme."
+echo "*If the terminal doesn't look updated close and restart the shell.*"
+
+update_theme "$NEW_THEME"
